use-cases/framework/agents/resources/agentcore-agent-handler/requirements.txt
bedrock-agentcore>=1.0.0
strands-agents>=1.0.0
boto3>=1.0.0
aws-lambda-powertools>=2.0.0
aws-opentelemetry-distro>=0.10.0
